refactor(model): Move some classes to the admin config service model
The classes to define assets for reports or report name mappings no
longer have counterparts in the ORT Server REST API. Therefore, there is
no reason for them to be located in the model module. Move them to the
model of the admin configuration.
