+++ services/admin-config/src/main/kotlin/ReporterConfig.kt
@@ -19,8 +19,69 @@
 
 package org.eclipse.apoapsis.ortserver.services.config
 
-import org.eclipse.apoapsis.ortserver.model.ReportNameMapping
-import org.eclipse.apoapsis.ortserver.model.ReporterAsset
+/**
+ * A class defining an asset (such as a font or an image) which is required to generate a report.
+ *
+ * The Reporter can be configured to download specific assets from the configuration before starting the report
+ * generation. That way it is ensured that files referenced from reporter templates are actually available locally at
+ * the expected relative paths.
+ *
+ * A [ReporterAsset] can be a single file or a directory. In the latter case, all files contained in the directory are
+ * downloaded.
+ */
+data class ReporterAsset(
+    /**
+     * The source path of this asset in the configuration. This path is passed to the configuration manager to download
+     * this asset.
+     */
+    val sourcePath: String,
+
+    /**
+     * A path (relative to the location of reporter template files) where this asset should be placed. Typically,
+     * assets are referenced via relative paths from reporter templates, e.g. _./images/logo.png_. Using this
+     * property, such a relative path can be specified. If it is `null`, the root folder of the reporter worker
+     * (which also contains the downloaded templates) is used.
+     */
+    val targetFolder: String? = null,
+
+    /**
+     * An optional name for the downloaded asset. This property can be used to rename the asset file or folder locally.
+     * If it is undefined, the original name (determined from the last path component of [sourcePath]) is used.
+     */
+    val targetName: String? = null
+)
+
+/**
+ * A class that can be used to configure the names under which reports are stored.
+ *
+ * Per default, the names of reports are determined by the reporters that produce these reports. This may not always
+ * be desired, especially if names are constructed dynamically based on some conventions. This class can be used to
+ * override the default names used by reporters. The reporter configuration contains a map that assigns each report
+ * type an optional instance of this class. The configuration contained in this instance is then used to generate the
+ * names for reports.
+ */
+data class ReportNameMapping(
+    /**
+     * A name prefix for constructing report names. The names of all reports generated by a reporter will start with
+     * this prefix and use the original file extensions. In case there are multiple reports, indices are added
+     * according to the further properties of this class.
+     */
+    val namePrefix: String,
+
+    /**
+     * The index to be used for the first report generated by a reporter. If there are multiple reports, their names
+     * are derived from [namePrefix] plus a sequential index starting with this value, such as "report-1.html",
+     * "report-2.html", etc.
+     */
+    val startIndex: Int = 1,
+
+    /**
+     * A flag that determines whether the name of reports should always contain an index. The flag has only an effect
+     * if there is only a single report. If it is *false*, the report is named by the [namePrefix] plus the original
+     * file extension. If it is *true*, the [startIndex] is appended to the [namePrefix].
+     */
+    val alwaysAppendIndex: Boolean = false
+)
 
 /**
  * A data class defining a report that is supported by an ORT Server instance. An instance specifies the concrete
