+++ services/admin-config/src/test/kotlin/AdminConfigServiceTest.kt
@@ -43,7 +43,6 @@ import org.eclipse.apoapsis.ortserver.config.ConfigException
 import org.eclipse.apoapsis.ortserver.config.ConfigManager
 import org.eclipse.apoapsis.ortserver.config.Context
 import org.eclipse.apoapsis.ortserver.config.Path
-import org.eclipse.apoapsis.ortserver.model.ReporterAsset
 import org.eclipse.apoapsis.ortserver.model.SourceCodeOrigin
 
 import org.ossreviewtoolkit.model.config.ScannerConfiguration
